Added new algorithms of generating random values. Added interfaces for new functionality.

+++ Diamond-Square/Interfaces/IHeightMapping.cs
@@ -2,6 +2,6 @@
 {
     public interface IHeightMapping
     {
-        float[,] GenerateHeightMap(IRandomGenerator randomGenerator);
+        int[,] GenerateHeightMap(IRandomGenerator randomGenerator);
     }
 }
